+++ app/controllers/variables_controller.rb
@@ -39,7 +39,7 @@ class VariablesController < ApplicationController
   # GET /projects/:project_id/variables
   def index
     @order = scrub_order(Variable, params[:order], 'variables.name')
-    variable_scope = viewable_variables.search(params[:search]).order(@order)
+    variable_scope = viewable_variables.search(params[:search], match_start: false).order(@order)
     variable_scope = variable_scope.where(user_id: params[:user_id]) if params[:user_id].present?
     variable_scope = variable_scope.where(variable_type: params[:variable_type]) if params[:variable_type].present?
     @variables = variable_scope.page(params[:page]).per(20)