+++ app/models/concerns/searchable.rb
@@ -9,12 +9,12 @@ module Searchable
     def self.search(arg, match_start: true)
       term = \
         if match_start
-          arg.to_s.downcase.gsub(/ |$/, '%')
+          arg.to_s.downcase.gsub(/ |$/, "%")
         else
-          arg.to_s.downcase.gsub(/^| |$/, '%')
+          arg.to_s.downcase.gsub(/^| |$/, "%")
         end
       terms = [term] * search_queries.count
-      where search_queries.join(' or '), *terms
+      where search_queries.join(" or "), *terms
     end
 
     def self.search_queries