+++ test/controllers/search_controller_test.rb
@@ -1,12 +1,12 @@
 # frozen_string_literal: true
 
-require 'test_helper'
+require "test_helper"
 
 # Tests to assure that search results are returned.
 class SearchControllerTest < ActionDispatch::IntegrationTest
-  test 'should get search' do
+  test "should get search" do
     login(users(:valid))
-    get search_url, params: { search: '' }
+    get search_url, params: { search: "" }
     assert_not_nil assigns(:subjects)
     assert_not_nil assigns(:projects)
     assert_not_nil assigns(:designs)
@@ -15,9 +15,9 @@ class SearchControllerTest < ActionDispatch::IntegrationTest
     assert_response :success
   end
 
-  test 'should get search and redirect' do
+  test "should get search and redirect to project" do
     login(users(:valid))
-    get search_url, params: { search: 'Project With One Design' }
+    get search_url, params: { search: "Project With One Design" }
     assert_not_nil assigns(:subjects)
     assert_not_nil assigns(:projects)
     assert_not_nil assigns(:designs)
@@ -27,14 +27,15 @@ class SearchControllerTest < ActionDispatch::IntegrationTest
     assert_redirected_to assigns(:objects).first
   end
 
-  test 'should get search typeahead' do
+  test "should get search and redirect to variable" do
     login(users(:valid))
-    get search_url(format: 'json'), params: { search: 'abc' }
+    get search_url, params: { search: "var_course_work" }
     assert_not_nil assigns(:subjects)
     assert_not_nil assigns(:projects)
     assert_not_nil assigns(:designs)
     assert_not_nil assigns(:variables)
     assert_not_nil assigns(:objects)
-    assert_response :success
+    assert_equal 1, assigns(:objects).size
+    assert_redirected_to [variables(:checkbox).project, variables(:checkbox)]
   end
 end